# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Picket API 명세서
  description: 공연 예매 플랫폼 Picket 백엔드 API 명세
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 로컬 개발 서버
  - url: https://api.picket.o-r.kr
    description: 운영 서버
...
paths:
...
  /reviews/register:
    post:
      summary: 리뷰 등록
      description: 공연 상품에 대한 리뷰를 등록한다.
      tags: [리뷰]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewRegisterRequest' }
      responses:
        '200':
          description: 등록 성공
          content:
            text/plain:
              schema: { type: string, example: "리뷰저장성공" }
        '400':
          description: 입력값 검증 실패
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /reviews/userReviewList:
    get:
      summary: 내가 쓴 리뷰 조회 (기간별)
      description: 로그인 유저가 특정 기간에 작성한 리뷰를 조회한다.
      tags: [리뷰]
      parameters:
        - in: query
          name: startDate
          required: true
          schema: { type: string }
        - in: query
          name: endDate
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      results:
                        type: array
                        items: { $ref: '#/components/schemas/ReviewListResult' }
  /reviews/listPaging:
    get:
      summary: 리뷰 목록 페이징 조회
      description: 특정 상품의 리뷰를 페이지·사이즈로 조회한다.
      tags: [리뷰]
      parameters:
        - in: query
          name: page
          required: true
          schema: { type: integer, minimum: 0, description: "0부터 시작" }
        - in: query
          name: size
          required: true
          schema: { type: integer, minimum: 1 }
        - in: query
          name: productId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      results: { $ref: '#/components/schemas/ReviewPageResult' }

  # ---------------- 문의 (QnA) ----------------
...
components:
  schemas:
...
    BaseResponse:
      type: object
      properties:
        success: { type: boolean }
        code: { type: integer, example: 20000 }
        message: { type: string, example: "요청에 성공하였습니다." }
        results:
          description: 응답 결과. 엔드포인트별 스키마 참조.
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        code: { type: integer, example: 40000 }
        message: { type: string, example: "잘못된 입력입니다." }
        results: { nullable: true }

    # ---------------- 인증 ----------------
...
    ReviewRegisterRequest:
      type: object
      required: [rating, comment, productId]
      properties:
        rating: { type: integer, minimum: 1, maximum: 5, example: 3 }
        comment: { type: string, example: "정말 좋은 공연이었습니다." }
        productId: { type: integer, format: int64, example: 1 }
    ReviewListResult:
      type: object
      description: 리뷰 단건 정보
    ReviewPageResult:
      type: object
      description: 페이징된 리뷰 목록

    # ---------------- 문의 ----------------
...